+++ tests/project/project/__init__.py
@@ -1,2 +1,3 @@
 __version__: str = "0.0.0"
 __version__ = "0.0.0"
+__version_tuple__ = (0, 0, 0)